requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
django-filter==22.1
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
Markdown==3.4.3
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
